if condition backend part 2

+++ app/Http/Controllers/GaleryController.php
@@ -19,11 +19,13 @@ class GaleryController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'filename' => 'required|image|mimes:png,jpeg,jpg',
             'list_location_id' => 'required'
         ]);
         
         if($request->hasFile('filename')) {
+            $request->validate([
+                'filename' => 'required|image|mimes:png,jpeg,jpg'
+            ]);
             $file = $request->file('filename');
             $filename = time() . '.' . $file->getClientOriginalExtension();
             $file->storeAs('public/dolankuy/', $filename);
@@ -42,34 +44,49 @@ class GaleryController extends Controller
     public function show($id)
     {
         $currentGalery = Galery::find($id);
-        return response()->json(compact('currentGalery'));
+        if(empty($currentGalery)){
+
+        } else {
+            return response()->json(compact('currentGalery'));
+        }
+        
     }
 
     
     public function update(Request $request, $id)
     {
         $galery = Galery::find($id);
+
+        if(empty($galery)){
+
+        } else {
         
-        $request->validate([
-            'filename' => 'required|image|mimes:png,jpeg,jpg',
-            'list_location_id' => 'required'
-        ]);
+            $request->validate([
+                'list_location_id' => 'required'
+            ]);
 
-        if($request->hasFile('filename')) {
-            Storage::delete('/public/dolankuy/' . $galery->filename);
-            $file = $request->file('filename');
-            $filename = time() . '.' . $file->getClientOriginalExtension();
-            $file->storeAs('public/dolankuy/', $filename);  
-        }else{
-            $filename=$request->filename;
+
+            if($request->hasFile('filename')) {
+
+                $request->validate([
+                    'filename' => 'required|image|mimes:png,jpeg,jpg'
+                ]);
+                
+                $file = $request->file('filename');
+                $filename = time() . '.' . $file->getClientOriginalExtension();
+                $file->storeAs('public/dolankuy/', $filename);  
+                Storage::delete('/public/dolankuy/' . $galery->filename);
+            }else{
+                $filename=$request->filename;
+            }
+
+            $galery->update([
+                'list_location_id' => $request->list_location_id,
+                'filename' => $filename,
+            ]);
+
+            return response()->json($galery);
         }
-
-        $galery->update([
-            'list_location_id' => $request->list_location_id,
-            'filename' => $filename,
-        ]);
-
-        return response()->json($galery);
 
     }
 
@@ -77,10 +94,13 @@ class GaleryController extends Controller
     public function destroy(Request $request, $id)
     {
         $galery = Galery::find($id);
-        if($request->hasFile('image')) {
+        if(empty($galery)){
+
+        }else {
             Storage::delete('/public/dolankuy/' . $galery->filename);
+            $galery->delete();
+            return response()->json($galery);
         }
-        $galery->delete();
-        return response()->json($galery);
+        
     }
 }
